Move some generic structs out of isort
